Add files via upload
UML format för diagrams.net
UML bild av diagrammet

Note: We updated the UML by separating GasPanel och ControlPanel into their own classes. Since the represent Controllers and not the Model. They should not
be part of the CarView. Because the CarView represents the model (thus in the View package).

We should later move GasPanel and ControlPanel to the controller package. Since they represent visual representations of buttons and actions. But doesn't represent the model